feat: add list accounts

+++ src/factories/AccountFactory.java
@@ -6,5 +6,5 @@ import model.AccountType;
 import java.time.LocalDate;
 
 public interface AccountFactory {
-    Account create(String ownerName, int number, int digit, int funds, LocalDate openingDate, AccountType accountType);
+    Account create(String ownerName, int number, int digit, double funds, LocalDate openingDate, AccountType accountType);
 }
