+++ src/factories/CurrentAccountFactory.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 
 public class CurrentAccountFactory implements AccountFactory {
     @Override
-    public Account create(String ownerName, int number, int digit, int funds, LocalDate openingDate, AccountType accountType) {
+    public Account create(String ownerName, int number, int digit, double funds, LocalDate openingDate, AccountType accountType) {
         return new CurrentAccount(ownerName, number, digit, funds, openingDate, accountType);
     }
 }
